requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.12
django-heroku==0.3.1
Pillow==9.1.0
psycopg2-binary==2.8.3
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==6.0.0
gunicorn==20.1.0
djangorestframework==3.13.1
